finishing up security tokens throughout, fixing subnav in IE 7
git-svn-id: http://svn.concrete5.org/svn/concrete5@334 b0551a0c-1e16-4222-a7d5-975db1aca215

+++ trunk/web/concrete/tools/al_delete.php
@@ -3,22 +3,28 @@
 defined('C5_EXECUTE') or die(_("Access Denied."));
 
 $resp = array();
-if (isset($_REQUEST['bID'])) {
-	$c = Page::getByPath('/dashboard/mediabrowser');
-	$cp = new Permissions($c);
-	if ($cp->canRead()) {
-		$b = Block::getByID($_REQUEST['bID']);
-		if (is_object($b)) {
-			$b->delete();
-			$resp['error'] = 0;
+$valt = Loader::helper('validation/token');
+if ($valt->validate('delete_file')) {
+	if (isset($_REQUEST['bID'])) {
+		$c = Page::getByPath('/dashboard/mediabrowser');
+		$cp = new Permissions($c);
+		if ($cp->canRead()) {
+			$b = Block::getByID($_REQUEST['bID']);
+			if (is_object($b)) {
+				$b->delete();
+				$resp['error'] = 0;
+			} else {
+				$resp['error'] = 1;
+				$resp['message'] = t('Could not find block.');
+			}
 		} else {
 			$resp['error'] = 1;
-			$resp['message'] = t('Could not find block.');
+			$resp['message'] = t('You do not have permission to remove that block.');
 		}
-	} else {
-		$resp['error'] = 1;
-		$resp['message'] = t('You do not have permission to remove that block.');
 	}
+} else {
+	$resp['error'] = 1;
+	$resp['message'] = $valt->getErrorMessage();
 }
 $js = Loader::helper('json');
 print $js->encode($resp);
